# FilmApi/DTOs/Movie/ReadMovieDto.cs
namespace FilmApi.DTOs.Movie
{
    public class ReadMovieDto
    {
        public string Title { get; set; }

        public string Genre { get; set; }
        public int ReleaseYear { get; set; }
        public string Director { get; set; }

        public string Picture { get; set; }

        public string Trailer { get; set; }

        public int? FranchiseId { get; set; }
    }
}